+++ cognite/experimental/_api/functions.py
@@ -6,7 +6,15 @@ from zipfile import ZipFile
 
 from cognite.client import utils
 from cognite.client._api_client import APIClient
-from cognite.experimental.data_classes import Function, FunctionCall, FunctionCallList, FunctionCallLog, FunctionList
+from cognite.experimental.data_classes import (
+    Function,
+    FunctionCall,
+    FunctionCallList,
+    FunctionCallLog,
+    FunctionList,
+    FunctionSchedule,
+    FunctionSchedulesList,
+)
 
 
 class FunctionsAPI(APIClient):
@@ -16,6 +24,7 @@ class FunctionsAPI(APIClient):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self.calls = FunctionCallsAPI(*args, **kwargs)
+        self.schedules = FunctionSchedulesAPI(*args, **kwargs)
 
     def create(
         self,
@@ -110,7 +119,7 @@ class FunctionsAPI(APIClient):
 
         Returns:
             FunctionList: List of functions
-        
+
         Example:
 
             List functions::
@@ -296,7 +305,7 @@ class FunctionCallsAPI(APIClient):
 
         Returns:
             FunctionCallList: List of function calls
-        
+
         Examples:
 
             List function calls::
@@ -332,7 +341,7 @@ class FunctionCallsAPI(APIClient):
 
         Returns:
             FunctionCall: Function call.
-        
+
         Examples:
 
             Retrieve function call by id::
@@ -368,7 +377,7 @@ class FunctionCallsAPI(APIClient):
 
         Returns:
             FunctionCallLog: Log for the function call.
-        
+
         Examples:
 
             Retrieve function call logs by call ID::
@@ -391,3 +400,94 @@ class FunctionCallsAPI(APIClient):
         url = f"/functions/{function_id}/calls/{call_id}/logs"
         res = self._get(url)
         return FunctionCallLog._load(res.json()["items"])
+
+
+class FunctionSchedulesAPI(APIClient):
+    def list(self) -> FunctionSchedulesList:
+        """`List all schedules associated with a specific project. <https://docs.cognite.com/api/playground/#operation/get-api-playground-projects-project-functions-schedules>`_
+
+        Returns:
+            FunctionSchedulesList: List of function schedules
+
+        Examples:
+
+            List function schedules::
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>> schedules = c.functions.schedules.list()
+
+        """
+        url = f"/functions/schedules"
+        res = self._get(url)
+        return FunctionSchedulesList._load(res.json()["items"])
+
+    def create(
+        self,
+        name: str,
+        function_external_id: str,
+        cron_expression: str,
+        description: str = "",
+        data: Optional[Dict] = None,
+    ) -> FunctionSchedule:
+        """`Create a schedule associated with a specific project. <https://docs.cognite.com/api/playground/#operation/post-api-playground-projects-project-functions-schedules>`_
+
+        Args:
+            name (str): Name of the schedule.
+            function_external_id (str): External id of the function.
+            description (str): Description of the schedule.
+            cron_expression (str): Cron expression.
+            data (optional, Dict): Data to be passed to the scheduled run.
+
+        Returns:
+            FunctionSchedule: Created function schedule.
+
+        Examples:
+
+            Create function schedule::
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>> schedule = c.functions.schedules.create(
+                    name= "My schedule",
+                    function_external_id="my-external-id",
+                    cron_expression="*/5 * * * *",
+                    description="This schedule does magic stuff.")
+
+        """
+        json = {
+            "items": [
+                {
+                    "name": name,
+                    "description": description,
+                    "functionExternalId": function_external_id,
+                    "cronExpression": cron_expression,
+                    "data": data,
+                }
+            ]
+        }
+        url = f"/functions/schedules"
+        res = self._post(url, json=json)
+        return FunctionSchedule._load(res.json()["items"][0])
+
+    def delete(self, id: int) -> None:
+        """`Delete a schedule associated with a specific project. <https://docs.cognite.com/api/playground/#operation/post-api-playground-projects-project-functions-schedules-delete>`_
+
+        Args:
+            id (int): Id of the schedule
+
+        Returns:
+            None
+
+        Examples:
+
+            Delete function schedule::
+
+                >>> from cognite.experimental import CogniteClient
+                >>> c = CogniteClient()
+                >>> c.functions.schedules.delete(id = 123)
+
+        """
+        json = {"items": [{"id": id,}]}
+        url = f"/functions/schedules/delete"
+        self._post(url, json=json)